requirements.txt
asgiref
attrs
Automat
bcrypt
blinker
certifi
chardet
click
colorama
constantly
crispy-bootstrap5
cryptography
defer
distlib
distro
distro-info
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
dj-database-url
duplicity
fasteners
filelock
future
httplib2
hyperlink
idna
importlib-metadata
incremental
jeepney
keyring
lockfile
markdown
Mako
MarkupSafe
monotonic
more-itertools
MouseInfo
netifaces
oauthlib
paramiko
pexpect
pillow
pipenv
psycopg2-binary
ptyprocess
pyasn1
pyasn1-modules
PyAutoGUI
pycairo
PyGetWindow
Pygments
PyHamcrest
PyJWT
pymacaroons
PyMsgBox
PyNaCl
pyOpenSSL
pyparsing
pyperclip
PyRect
pyRFC3339
PyScreeze
python-dateutil
python-debian
python3-xlib
pytweening
pytz
pyxdg
PyYAML
rencode
requests
SecretStorage
service-identity
setproctitle
six
sqlparse
Twisted
typing_extensions
urllib3
virtualenv
zipp
zope.interface